equity_quotes/requirements.txt
quixstreams
python-dotenv
websocket-client
quixplus==4.1.1